requirements.txt
flask==2.3.3
flask-socketio==5.3.6
psutil==5.9.5
PySocks==1.7.1
